refactor(sonar): apply code quality and security fixes to pass quality gate
- Reduced cognitive complexity via helper methods
- Replaced unsafe Optional.get() usages
- Sanitized logs to avoid user-controlled data exposure
- Extracted duplicated literals into constants
- Simplified loops and control flow
- Updated streams to use toList()
- Minor test cleanup (removed unused params)
- Preserved behavior and API contracts
- Build passes (mvn clean install)

+++ src/main/java/com/tourism/platform/service/impl/MatchingServiceImpl.java
@@ -100,7 +100,7 @@ public class MatchingServiceImpl implements MatchingService {
         );
         int interestPoints = computeInterestPoints(normalizedInterests, userInterests);
 
-        double totalScore = destinationPoints + datePoints + interestPoints;
+        double totalScore = (double) destinationPoints + datePoints + interestPoints;
 
         return MatchResponseDto.builder()
                 .userId(plan.getUser().getId())
